app for float32_test added

+++ workspace/float32_test/app.cpp
@@ -1,13 +1,13 @@
 #include "app.h"
 #include "mros2.h"
-#include "std_msgs/msg/int8.hpp"
+#include "std_msgs/msg/float32.hpp"
 
 #include "stm32f7xx_nucleo_144.h"
 
 mros2::Subscriber sub;
 mros2::Publisher pub;
 
-void userCallback(std_msgs::msg::Int8 *msg)
+void userCallback(std_msgs::msg::Float32 *msg)
 {
   MROS2_INFO("subscribed msg: %d", msg->data);
   MROS2_INFO("publishing msg: %d", msg->data);
@@ -23,9 +23,9 @@ int main(int argc, char * argv[])
   BSP_LED_Toggle(LED1);
 
   mros2::Node node = mros2::Node::create_node("mros2_node");
-  pub = node.create_publisher<std_msgs::msg::Int8>("to_linux", 10);
+  pub = node.create_publisher<std_msgs::msg::Float32>("to_linux", 10);
   sub = node.create_subscription("to_stm", 10, userCallback);
-  std_msgs::msg::Int8 msg;
+  std_msgs::msg::Float32 msg;
 
   MROS2_INFO("ready to pub/sub message");
   mros2::spin();
